requirements.txt
fastapi
uvicorn
transformers
torch
requests
python-logstash-async
hvac